Reworked L movement script and renamed to MultiPointMovement
